AllowedMethodsHandler works even when list of methods is not sorted
Previous version required the input slice of strings to be sorted,
which was not obvious by the name of the function, and therefore,
would lead to confusing and erroneus behavior.

This version copies the function argument representing the slice of
strings, sorts the list, and uses that sorted list during the checks.