no longer panics concurrent map access when handler modifies http headers
When any request headers are to be logged, this must copy the
respective values before it creates a go routine to handle
request.  Otherwise, if this must later time out the invoked
request header before that returns, that handler might
concurrently try to alter request headers while this is reading
them to emit the log line.