responseWriter refactor
No longer needs to create a new response writer when handler times out
to protect against race between handler writing to underlying response
writer and this writing to it.

Removed some lock contention on common cases for responseWriter,
namely no longer need it for when the handler completes normally or
when the handler panics. Only need lock to protect response writers
public methods from being called when in the middle of responding for
a handler that has not responded before the context closed.